# Conflicts:
#	.pre-commit-config.yaml
#	setup.py
#	src/octoprint/plugins/pluginmanager/__init__.py
#	src/octoprint/server/api/printer.py
#	src/octoprint/server/api/printer_profiles.py
#	src/octoprint/server/api/settings.py
#	src/octoprint/util/commandline.py

+++ src/octoprint/plugins/logging/__init__.py
@@ -6,7 +6,7 @@ import os
 from os import scandir
 
 import yaml
-from flask import jsonify, make_response, request, url_for
+from flask import abort, jsonify, request, url_for
 from flask_babel import gettext
 from werkzeug.exceptions import BadRequest
 from werkzeug.utils import secure_filename
@@ -17,6 +17,7 @@ from octoprint.access.permissions import Permissions
 from octoprint.server import NO_CONTENT
 from octoprint.server.util.flask import no_firstrun_access, redirect_to_tornado
 from octoprint.settings import settings
+from octoprint.util import is_hidden_path
 
 
 class LoggingPlugin(
@@ -81,8 +82,12 @@ class LoggingPlugin(
     @Permissions.PLUGIN_LOGGING_MANAGE.require(403)
     def delete_log(self, filename):
         secure = os.path.join(settings().getBaseFolder("logs"), secure_filename(filename))
-        if not os.path.exists(secure):
-            return make_response("File not found: %s" % filename, 404)
+        if (
+            not os.path.exists(secure)
+            or is_hidden_path(secure)
+            or not filename.endswith(".log")
+        ):
+            abort(404)
 
         os.remove(secure)
 
@@ -107,15 +112,15 @@ class LoggingPlugin(
     @Permissions.PLUGIN_LOGGING_MANAGE.require(403)
     def set_logging_levels_api(self):
         if "application/json" not in request.headers["Content-Type"]:
-            return make_response("Expected content-type JSON", 400)
+            abort(400, description="Expected content-type JSON")
 
         try:
             json_data = request.json
         except BadRequest:
-            return make_response("Malformed JSON body in request", 400)
+            abort(400, description="Malformed JSON body in request")
 
         if not isinstance(json_data, dict):
-            return make_response("Invalid log level configuration", 400)
+            abort(400, description="Invalid log level configuration")
 
         # TODO validate further
 
@@ -135,6 +140,12 @@ class LoggingPlugin(
         files = []
         basedir = settings().getBaseFolder("logs", check_writable=False)
         for entry in scandir(basedir):
+            if is_hidden_path(entry.path) or not entry.name.endswith(".log"):
+                continue
+
+            if not entry.is_file():
+                continue
+
             files.append(
                 {
                     "name": entry.name,
